+++ src/Data/Geo/Jord.hs
@@ -15,28 +15,28 @@
 --
 module Data.Geo.Jord
     ( module Data.Geo.Jord.Angle
-    , module Data.Geo.Jord.Ecef
     , module Data.Geo.Jord.Ellipsoid
     , module Data.Geo.Jord.Eval
-    , module Data.Geo.Jord.GreatCircle
-    , module Data.Geo.Jord.HorizontalPosition
     , module Data.Geo.Jord.LatLong
     , module Data.Geo.Jord.Length
+    , module Data.Geo.Jord.NVector
+    , module Data.Geo.Jord.Position
     , module Data.Geo.Jord.Quantity
-    , module Data.Geo.Jord.Vector3d
+    , module Data.Geo.Jord.Spherical.Geodetics
+    , module Data.Geo.Jord.Spherical.GreatCircle
     , jordVersion
     ) where
 
 import Data.Geo.Jord.Angle
-import Data.Geo.Jord.Ecef
 import Data.Geo.Jord.Ellipsoid
 import Data.Geo.Jord.Eval
-import Data.Geo.Jord.GreatCircle
-import Data.Geo.Jord.HorizontalPosition
 import Data.Geo.Jord.LatLong
 import Data.Geo.Jord.Length
+import Data.Geo.Jord.NVector
+import Data.Geo.Jord.Position
 import Data.Geo.Jord.Quantity
-import Data.Geo.Jord.Vector3d
+import Data.Geo.Jord.Spherical.Geodetics
+import Data.Geo.Jord.Spherical.GreatCircle
 
 -- | version.
 jordVersion :: String