+++ src/Data/Geo/Jord/Spherical/Geodetics.hs
@@ -1,24 +1,18 @@
 -- |
--- Module:      Data.Geo.Jord.HorizontalPosition
+-- Module:      Data.Geo.Jord.Spherical.Geodetics
 -- Copyright:   (c) 2018 Cedric Liegeois
 -- License:     BSD3
 -- Maintainer:  Cedric Liegeois <ofmooseandmen@yahoo.fr>
 -- Stability:   experimental
 -- Portability: portable
 --
--- Types and functions for working with horizontal positions.
+-- Geodetics calculations on horizontal positions assuming a spherical earth model.
 --
 -- All functions are implemented using the vector-based approached described in
 -- <http://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf Gade, K. (2010). A Non-singular Horizontal Position Representation>
 --
--- This module assumes a spherical earth.
---
-module Data.Geo.Jord.HorizontalPosition
-    (
-    -- * The 'HorizontalPosition' type
-      HorizontalPosition(..)
-    -- * Geodetic calculations
-    , angularDistance
+module Data.Geo.Jord.Spherical.Geodetics
+    ( angularDistance
     , antipode
     , destination
     , destination'
@@ -29,81 +23,49 @@ module Data.Geo.Jord.HorizontalPosition
     , interpolate
     , isInside
     , mean
-    -- * Remarkable positions.
-    , northPole
-    , southPole
     ) where
 
 import Data.Geo.Jord.Angle
 import Data.Geo.Jord.Ellipsoid
 import Data.Geo.Jord.LatLong
 import Data.Geo.Jord.Length
+import Data.Geo.Jord.NVector
+import Data.Geo.Jord.Position (Geodetic2D(..), northPole)
 import Data.Geo.Jord.Quantity
-import Data.Geo.Jord.Vector3d
 import Data.List (subsequences)
 import Prelude hiding (fail)
 
--- | The 'HorizontalPosition' class defines 2 functions to convert a Horizontal Position to and from a n-vector.
--- All functions in this module first convert 'HorizontalPosition' to a n-vector and any resulting n-vector back
--- to a 'HorizontalPosition'. This allows the call site to pass either n-vector or another 'HorizontalPosition' instance
--- and to get back the same class instance.
-class (Eq a) =>
-      HorizontalPosition a where
-    -- | Converts a 'Vector3d' into 'HorizontalPosition' instance.
-    fromNVector :: Vector3d -> a
-    -- | Converts the 'HorizontalPosition' instance into a 'Vector3d'.
-    toNVector :: a -> Vector3d
-
--- | 'LatLong' to/from 'Vector3d'.
-instance HorizontalPosition LatLong where
-    fromNVector v = latLong lat lon
-      where
-        lat = atan2' (z v) (sqrt (x v * x v + y v * y v))
-        lon = atan2' (y v) (x v)
-    toNVector g = Vector3d x' y' z'
-      where
-        lat = latitude g
-        lon = longitude g
-        cl = cos' lat
-        x' = cl * cos' lon
-        y' = cl * sin' lon
-        z' = sin' lat
-
--- | Identity.
-instance HorizontalPosition Vector3d where
-    fromNVector v = v
-    toNVector v = v
-
--- | Angle between the two given 'HorizontalPosition's.
+-- | @angularDistance p1 p2 n@ computes the angle between the horizontal positions @p1@ and @p2@.
 -- If @n@ is 'Nothing', the angle is always in [0..180], otherwise it is in [-180, +180],
 -- signed + if @p1@ is clockwise looking along @n@, - in opposite direction.
-angularDistance :: (HorizontalPosition a) => a -> a -> Maybe a -> Angle
+angularDistance :: (Geodetic2D a) => a -> a -> Maybe a -> Angle
 angularDistance p1 p2 n = angularDistance' (toNVector p1) (toNVector p2) (fmap toNVector n)
 
--- | Angle between the two given 'Vector3d's.
+-- | Angle between the two given n-vectors.
 -- If @n@ is 'Nothing', the angle is always in [0..180], otherwise it is in [-180, +180],
 -- signed + if @v1@ is clockwise looking along @n@, - in opposite direction.
-angularDistance' :: Vector3d -> Vector3d -> Maybe Vector3d -> Angle
+angularDistance' :: NVector -> NVector -> Maybe NVector -> Angle
 angularDistance' v1 v2 n = atan2' sinO cosO
   where
     sign = maybe 1 (signum . dot (cross v1 v2)) n
     sinO = sign * norm (cross v1 v2)
     cosO = dot v1 v2
 
--- | Returns the antipodal 'HorizontalPosition' of the given 'HorizontalPosition' - i.e. the HorizontalPosition on the surface
--- of the Earth which is diametrically opposite to the given HorizontalPosition.
-antipode :: (HorizontalPosition a) => a -> a
+-- | @antipode p@ computes the antipodal horizontal position of @p@:
+-- the horizontal position on the surface of the Earth which is diametrically opposite to @p@.
+antipode :: (Geodetic2D a) => a -> a
 antipode p = fromNVector (scale (toNVector p) (-1.0))
 
 -- | 'destination'' using the mean radius of the WGS84 reference ellipsoid.
-destination :: (HorizontalPosition a) => a -> Angle -> Length -> a
+destination :: (Geodetic2D a) => a -> Angle -> Length -> a
 destination p b d = destination' p b d meanEarthRadius
 
--- | Computes the destination 'HorizontalPosition' from the given 'HorizontalPosition' having travelled the given distance on the
--- given initial bearing (bearing will normally vary before destination is reached) and using the given earth radius.
+-- | @destination' p b d r@ computes the destination horizontal position from position @p@ having
+-- travelled the distance @d@ on the initial bearing @b@ (bearing will normally vary
+-- before destination is reached) and using the earth radius @r@.
 --
 -- This is known as the direct geodetic problem.
-destination' :: (HorizontalPosition a) => a -> Angle -> Length -> Length -> a
+destination' :: (Geodetic2D a) => a -> Angle -> Length -> Length -> a
 destination' p b d r
     | isZero d = p
     | otherwise = fromNVector (add (scale v (cos' ta)) (scale de (sin' ta)))
@@ -115,23 +77,23 @@ destination' p b d r
     de = add (scale nd (cos' b)) (scale ed (sin' b)) -- unit vector in the direction of the azimuth
 
 -- | 'distance'' using the mean radius of the WGS84 reference ellipsoid.
-distance :: (HorizontalPosition a) => a -> a -> Length
+distance :: (Geodetic2D a) => a -> a -> Length
 distance p1 p2 = distance' p1 p2 meanEarthRadius
 
--- | Computes the surface distance (length of geodesic) in 'Meters' assuming a
--- spherical Earth between the two given 'HorizontalPosition's and using the given earth radius.
-distance' :: (HorizontalPosition a) => a -> a -> Length -> Length
+-- | @distance' p1 p2@ computes the surface distance (length of geodesic) between the positions @p1@ and @p2@
+-- and using the earth radius @r@.
+distance' :: (Geodetic2D a) => a -> a -> Length -> Length
 distance' p1 p2 = arcLength (angularDistance p1 p2 Nothing)
 
--- | Computes the final bearing arriving at given destination  @p2@ 'HorizontalPosition' from given 'HorizontalPosition' @p1@.
---  the final bearing will differ from the 'initialBearing' by varying degrees according to distance and latitude.
--- Returns 180 if both HorizontalPosition are equals.
-finalBearing :: (HorizontalPosition a) => a -> a -> Angle
+-- | @finalBearing p1 p2@ computes the final bearing arriving at @p2@ from @p1@ in compass angle.
+--  The final bearing will differ from the 'initialBearing' by varying degrees according to distance and latitude.
+-- Returns 180 if both horizontal positions are equals.
+finalBearing :: (Geodetic2D a) => a -> a -> Angle
 finalBearing p1 p2 = normalise (initialBearing p2 p1) (decimalDegrees 180)
 
--- | Computes the initial bearing from given @p1@ 'HorizontalPosition' to given @p2@ 'HorizontalPosition', in compass degrees.
--- Returns 0 if both HorizontalPosition are equals.
-initialBearing :: (HorizontalPosition a) => a -> a -> Angle
+-- | @initialBearing p1 p2@ computes the initial bearing from @p1@ to @p2@ in compass angle.
+-- Returns 0 if both horizontal positions are equals.
+initialBearing :: (Geodetic2D a) => a -> a -> Angle
 initialBearing p1 p2 = normalise (angularDistance' gc1 gc2 (Just v1)) (decimalDegrees 360)
   where
     v1 = toNVector p1
@@ -139,7 +101,7 @@ initialBearing p1 p2 = normalise (angularDistance' gc1 gc2 (Just v1)) (decimalDe
     gc1 = cross v1 v2 -- great circle through p1 & p2
     gc2 = cross v1 northPole -- great circle through p1 & north pole
 
--- | Computes the 'HorizontalPosition' at given fraction @f@ between the two given 'HorizontalPosition's @p0@ and @p1@.
+-- | @interpolate p0 p1 f# computes the horizontal position at fraction @f@ between the @p0@ and @p1@.
 --
 -- Special conditions:
 --
@@ -150,7 +112,7 @@ initialBearing p1 p2 = normalise (angularDistance' gc1 gc2 (Just v1)) (decimalDe
 --
 -- 'error's if @f < 0 || f > 1.0@
 --
-interpolate :: (HorizontalPosition a) => a -> a -> Double -> a
+interpolate :: (Geodetic2D a) => a -> a -> Double -> a
 interpolate p0 p1 f
     | f < 0 || f > 1 = error ("fraction must be in range [0..1], was " ++ show f)
     | f == 0 = p0
@@ -160,15 +122,15 @@ interpolate p0 p1 f
     v0 = toNVector p0
     v1 = toNVector p1
 
--- | Determines whether the given 'HorizontalPosition' is inside the polygon defined by the given list of 'HorizontalPosition's.
+-- | @isInside p ps@ determines whether the @p@ is inside the polygon defined by the list of positions @ps@.
 -- The polygon is closed if needed (i.e. if @head ps /= last ps@).
 --
 -- Uses the angle summation test: on a sphere, due to spherical excess, enclosed point angles
 -- will sum to less than 360°, and exterior point angles will be small but non-zero.
 --
--- Always returns 'False' if HorizontalPositions does not at least defines a triangle.
+-- Always returns 'False' if @ps@ does not at least defines a triangle.
 --
-isInside :: (Eq a, HorizontalPosition a) => a -> [a] -> Bool
+isInside :: (Eq a, Geodetic2D a) => a -> [a] -> Bool
 isInside p ps
     | null ps = False
     | head ps == last ps = isInside p (init ps)
@@ -182,14 +144,14 @@ isInside p ps
     es = egdes (map (sub v . toNVector) ps)
 
 -- | [p1, p2, p3, p4] to [(p1, p2), (p2, p3), (p3, p4), (p4, p1)]
-egdes :: [Vector3d] -> [(Vector3d, Vector3d)]
+egdes :: [NVector] -> [(NVector, NVector)]
 egdes ps = zip ps ps'
   where
     ps' = tail ps ++ [head ps]
 
--- | Computes the geographic mean 'HorizontalPosition' of the given 'HorizontalPosition's if it is defined.
+-- | @mean ps@ computes the mean geographic horitzontal position given list of position @ps@, if it is defined.
 --
--- The geographic mean is not defined for the antipodals HorizontalPositions (since they
+-- The geographic mean is not defined for antipodals position (since they
 -- cancel each other).
 --
 -- Special conditions:
@@ -201,7 +163,7 @@ egdes ps = zip ps ps'
 --     mean [p1, .., antipode p1] == Nothing
 -- @
 --
-mean :: (HorizontalPosition a) => [a] -> Maybe a
+mean :: (Geodetic2D a) => [a] -> Maybe a
 mean [] = Nothing
 mean [p] = Just p
 mean ps =
@@ -215,14 +177,6 @@ mean ps =
         filter
             (\t -> (fromNVector (antipode (head t)) :: LatLong) == (fromNVector (last t) :: LatLong))
             ts
-
--- | 'HorizontalPosition' of the North Pole.
-northPole :: (HorizontalPosition a) => a
-northPole = fromNVector (Vector3d 0.0 0.0 1.0)
-
--- | 'HorizontalPosition' of the South Pole.
-southPole :: (HorizontalPosition a) => a
-southPole = fromNVector (Vector3d 0.0 0.0 (-1.0))
 
 -- | WGS84 ellipsoid mean radius.
 meanEarthRadius :: Length